Add delete user capability
This change adds the delete user capability.
A DB change is required. Run the following commands to migrate/upgrade the DB:
pip install Flask-Migrate
flask db init # optional if db already exists
flask db migrate -m "Initial migration."
flask db upgrade

A new delete button is added to the admin dashbord. When clicked, the new 'is_deleted' DB column will be set to true and the email will be automatically prefixed with DELETED_PREFIX.
To undelete a user, use the Edit button and uncheck the Deleted checkbox then save, the user will show up as undeleted and the DELETED_PREFIX will be automatically removed.

A new Purge Deleted Users button is added at the end of the users list in the admin dashboard. It will purge all users marked as deleted. An alert will show up to confirm the purge as it cannot be undone. Purging will remove the 'is_deleted' records from the DB.